Fixes typo in docs

+++ lib/mindwendel/brainstormings.ex
@@ -306,14 +306,13 @@ defmodule Mindwendel.Brainstormings do
 
   ## Examples
 
-      iex> add_like(1, 2)
+      iex> delete_like(1, 2)
       {:ok, %Idea{}}
 
   """
   def delete_like(idea_id, user_id) do
-    Repo.delete_all(
-      from like in Like, where: like.user_id == ^user_id and like.idea_id == ^idea_id
-    )
+    # we ignore the result, delete_all returns the count of deleted items. We'll reload and broadcast the idea either way:
+    Repo.delete_all(from like in Like, where: like.user_id == ^user_id and like.idea_id == ^idea_id)
 
     {:ok, get_idea!(idea_id)} |> broadcast(:idea_updated)
   end
